day1 final commit

+++ springapp/src/main/java/com/examly/springapp/model/Feedback.java
@@ -10,7 +10,7 @@ import jakarta.persistence.ManyToOne;
 public class Feedback {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int feedbackId;
+    long feedbackId;
     String feedbackText;
     String date;
     @ManyToOne
@@ -20,10 +20,10 @@ public class Feedback {
     @JoinColumn(name = "investmentId" ,nullable=true)
     Investment investment;
     String category;
-    public int getFeedbackId() {
+    public long getFeedbackId() {
         return feedbackId;
     }
-    public void setFeedbackId(int feedbackId) {
+    public void setFeedbackId(long feedbackId) {
         this.feedbackId = feedbackId;
     }
     public String getFeedbackText() {
